Fixed Program.cs and made a window

+++ ODEON_Event_Management/ApplicationLayer/DataBaseController.cs
@@ -82,7 +82,7 @@ namespace ApplicationLayer
                 command.Connection.Open();
                 SqlDataReader sqlDataReader = command.ExecuteReader();
                 sqlDataReader.Read();
-                afvikling.ID = sqlDataReader["DatoId"];
+                afvikling.ID = (int)sqlDataReader["DatoId"];
             }
         }
